correction for API

+++ app/Http/Controllers/BookController.php
@@ -4,14 +4,35 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Book;
+use Illuminate\Support\Facades\Auth;
 
 class BookController extends Controller
 {
     //
     public function index()
     {
-        $books = Book::all();
-        // return view with all books
+        try {
+            if (Auth::check()) {
+
+                $books = Book::all()->toArray();
+                $response = [
+                    'success' => true,
+                    'books' => $books,
+                ];
+                return response()->json($response, 200);
+            }
+            $response = [
+                'success' => false,
+                'message' => 'Usuário nao autenticado',
+            ];
+            return response()->json($response, 401);
+        } catch (\Throwable $th) {
+            $response = [
+                'success' => false,
+                'message' => $th->getMessage(),
+            ];
+            return response()->json($response, 500);
+        }
     }
 
     public function create(Request $request)
@@ -22,17 +43,39 @@ class BookController extends Controller
             'ISBN' => 'required|numeric|unique:books',
             'value' => 'required|numeric',
         ]);
+        try {
+            //code...
+            // create the new book in the database
+            if(Auth::check()){
 
-        // create the new book in the database
-        $book = Book::create([
-            'status' => 'active',
-            'name' => $request->name,
-            'ISBN' => $request->ISBN,
-            'value' => $request->value,
-        ]);
+                $book = Book::create([
+                    'status' => 'active',
+                    'name' => $request->name,
+                    'ISBN' => $request->ISBN,
+                    'value' => $request->value,
+                ]);
 
-        // redirect the user to the books index page with a success message
-        return redirect()->route('books.index')->with('success', 'Book created successfully.');
+                $response = [
+                    'success' => true,
+                    'message' => 'livro criado com sucesso',
+                ];
+                return response()->json($response, 200);
+            }
+            $response = [
+                'success' => false,
+                'message' => 'Usuário nao autenticado',
+            ];
+            return response()->json($response, 401);
+        } catch (\Throwable $th) {
+            //throw $th;
+            $response = [
+                'success' => false,
+                'message' => $th->getMessage(),
+            ];
+            return response()->json($response, 500);
+        }
+
+
     }
 
     public function update(Request $request, $id)
@@ -47,29 +90,74 @@ class BookController extends Controller
             'ISBN' => 'required|numeric|unique:books,ISBN,'.$book->id,
             'value' => 'required|numeric',
         ]);
+        try {
+            //code...
+            if(Auth::check()) {
+                // update the book in the database
+                $book->fill([
+                    'status' => $request->status,
+                    'name' => $request->name,
+                    'ISBN' => $request->ISBN,
+                    'value' => $request->value,
+                ])->save();
 
-        // update the book in the database
-        $book->fill([
-            'status' => $request->status,
-            'name' => $request->name,
-            'ISBN' => $request->ISBN,
-            'value' => $request->value,
-        ])->save();
+                if($book) {
+                    $response = [
+                        'success' => true,
+                        'message' => 'Livro atualizado com sucesso',
+                    ];
+                    return response()->json($response, 200);
+                }
+            }
 
-        // redirect the user to the books index page with a success message
-        return redirect()->route('books.index')->with('success', 'Book updated successfully.');
+            $response = [
+                'success' => false,
+                'message' => 'Usuário nao autenticado',
+            ];
+            return response()->json($response, 401);
+        } catch (\Throwable $th) {
+            //throw $th;
+            $response = [
+                'success' => false,
+                'message' => $th->getMessage(),
+            ];
+            return response()->json($response, 500);
+        }
+
     }
 
     public function delete($id)
     {
         // implement book deletion logic
         // find the book by ID
-        $book = Book::findOrFail($id);
+        try {
+            //code...
+            if(Auth::check()) {
 
-        // delete the book from the database
-        $book->delete();
+                $book = Book::findOrFail($id);
+                // delete the book from the database
+                $book->delete();
 
-        // redirect the user to the books index page with a success message
-        return redirect()->route('books.index')->with('success', 'Book deleted successfully.');
+                $response = [
+                    'success' => true,
+                    'message' => 'Livro removido com sucesso',
+                ];
+                return response()->json($response, 200);
+            }
+            $response = [
+                'success' => false,
+                'message' => 'Usuário nao autenticado',
+            ];
+            return response()->json($response, 401);
+        } catch (\Throwable $th) {
+            //throw $th;
+            $response = [
+                'success' => false,
+                'message' => $th->getMessage(),
+            ];
+            return response()->json($response, 500);
+        }
+
+
     }
 }
